% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.add_experiment}
\alias{deepblue.add_experiment}
\title{add_experiment}
\usage{
deepblue.add_experiment(name = NULL, genome = NULL,
  epigenetic_mark = NULL, sample = NULL, technique = NULL,
  project = NULL, description = NULL, data = NULL, format = NULL,
  extra_metadata = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (experiment name)}

\item{genome}{- A string (the target genome)}

\item{epigenetic_mark}{- A string (epigenetic mark of the experiment)}

\item{sample}{- A string (id of the used sample)}

\item{technique}{- A string (technique used by this experiment)}

\item{project}{- A string (the project name)}

\item{description}{- A string (description of the experiment)}

\item{data}{- A string (the BED formated data)}

\item{format}{- A string (format of the provided data)}

\item{extra_metadata}{- A struct (additional metadata)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Inserts a new experiment with the given parameters.

id of the newly inserted experiment) or ('error', error_message)
}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue.faceting_experiments}},
  \code{\link{deepblue.list_experiments}},
  \code{\link{deepblue.list_recent_experiments}},
  \code{\link{deepblue.list_similar_experiments}}
}

