% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.add_genome}
\alias{deepblue.add_genome}
\title{add_genome}
\usage{
deepblue.add_genome(name = NULL, description = NULL, data = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (genome name)}

\item{description}{- A string (description of the genome)}

\item{data}{- A string (genome data)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Inserts a new genome with the given parameters.

id of the newly inserted genome) or ('error', error_message)
}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.chromosomes}},
  \code{\link{deepblue.list_genomes}},
  \code{\link{deepblue.list_similar_genomes}},
  \code{\link{deepblue.upload_chromosome}}
}

