% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.add_sample_from_gsm}
\alias{deepblue.add_sample_from_gsm}
\title{add_sample_from_gsm}
\usage{
deepblue.add_sample_from_gsm(name = NULL, gsm_id = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (biosource name)}

\item{gsm_id}{- A string (GSM ID)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Import sample from an existing GSM identifier.

id of the newly inserted sample) or ('error', error_message)
}
\seealso{
Other Inserting and listing samples: \code{\link{deepblue.add_sample}},
  \code{\link{deepblue.list_samples}}
}

