% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.cancel_request}
\alias{deepblue.cancel_request}
\title{cancel_request}
\usage{
deepblue.cancel_request(id = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{id}{- A string (Request ID to be canceled, stopped or removed.)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Stop, cancel, and remove request data. Its data will be remove if the request did finish.

ID of the canceled request) or ('error', error_message)
}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.get_state}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.remove}},
  \code{\link{deepblue.search}}
}

