% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.change_extra_metadata}
\alias{deepblue.change_extra_metadata}
\title{change_extra_metadata}
\usage{
deepblue.change_extra_metadata(id = NULL, extra_metadata_key = NULL,
  extra_metadata_value = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{id}{- A string (id of the data)}

\item{extra_metadata_key}{- A string (extra_metadata key)}

\item{extra_metadata_value}{- A string (extra_metadata key (empty for delete this key))}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Change the extra metadata content for experiments, annotations, biosources, and samples.

id of the modified data) or ('error', error_message)
}
\seealso{
Other Operations that modify the data content: \code{\link{deepblue.clone_dataset}}
}

