% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.chromosomes}
\alias{deepblue.chromosomes}
\title{chromosomes}
\usage{
deepblue.chromosomes(genome = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{genome}{- A string (the target genome)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
List all chromosomes of a given genome.

A list containing all chromosomes, with theirs names and sizes) or ('error', error_message)
}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.add_genome}},
  \code{\link{deepblue.list_genomes}},
  \code{\link{deepblue.list_similar_genomes}},
  \code{\link{deepblue.upload_chromosome}}
}

