% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.echo}
\alias{deepblue.echo}
\title{echo}
\usage{
deepblue.echo(user_key = deepblue.USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Echos the server's version.

echo message including version) or ('error', error_message)
}
\seealso{
Other Checking DeepBlue status: \code{\link{deepblue.commands}}
}

