% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.extract_ids}
\alias{deepblue.extract_ids}
\title{extract_ids}
\usage{
deepblue.extract_ids(list = NULL)
}
\arguments{
\item{list}{- A array (list of lists of IDs and Names)}
}
\value{
a vector with  ('okay',
}
\description{
Extract the names from a list of ID and Names.

list containing the extracted IDs) or ('error', error_message)
}
\seealso{
Other Utilities for connecting operations: \code{\link{deepblue.extract_names}}
}

