% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.find_pattern}
\alias{deepblue.find_pattern}
\title{find_pattern}
\usage{
deepblue.find_pattern(pattern = NULL, genome = NULL, overlap = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{pattern}{- A string (pattern (regular expression))}

\item{genome}{- A string (the target genome)}

\item{overlap}{- A boolean (if the matching should do overlap search)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Process an annotation that will contain all genomic positions from the given pattern.

id of the annotation that contains the positions of the given pattern) or ('error', error_message)
}
\seealso{
Other Inserting and listing annotations: \code{\link{deepblue.add_annotation}},
  \code{\link{deepblue.list_annotations}}
}

