% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.is_biosource}
\alias{deepblue.is_biosource}
\title{is_biosource}
\usage{
deepblue.is_biosource(biosource_name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{biosource_name}{- A string (Name of the biosource)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Return information for the given biosource name.

A string containing the biosource name) or ('error', error_message)
}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.get_state}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.list_in_use}},
  \code{\link{deepblue.remove}},
  \code{\link{deepblue.search}}
}

