% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_genomes}
\alias{deepblue.list_genomes}
\title{list_genomes}
\usage{
deepblue.list_genomes(user_key = deepblue.USER_KEY)
}
\arguments{
\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all existing genomes.

genome names) or ('error', error_message)
}
\seealso{
Other Inserting and listing genomes: \code{\link{deepblue.add_genome}},
  \code{\link{deepblue.chromosomes}},
  \code{\link{deepblue.list_similar_genomes}},
  \code{\link{deepblue.upload_chromosome}}
}

