% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_in_use}
\alias{deepblue.list_in_use}
\title{list_in_use}
\usage{
deepblue.list_in_use(controlled_vocabulary = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{controlled_vocabulary}{- A string (id of the data)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all terms from the given controlled vocabulary that are used.

controlled_vocabulary terms with count) or ('error', error_message)
}
\seealso{
Other Commands for all types of data: \code{\link{deepblue.cancel_request}},
  \code{\link{deepblue.get_state}},
  \code{\link{deepblue.info}},
  \code{\link{deepblue.is_biosource}},
  \code{\link{deepblue.remove}},
  \code{\link{deepblue.search}}
}

