% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_recent_experiments}
\alias{deepblue.list_recent_experiments}
\title{list_recent_experiments}
\usage{
deepblue.list_recent_experiments(days = NULL, genome = NULL,
  epigenetic_mark = NULL, sample = NULL, technique = NULL,
  project = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{days}{- A double (maximum days ago the experiments were added)}

\item{genome}{- A string or a vector of string (the target genome)}

\item{epigenetic_mark}{- A string or a vector of string (name(s) of selected epigenetic mark(s))}

\item{sample}{- A string or a vector of string (id(s) of selected sample(s))}

\item{technique}{- A string or a vector of string (name(s) of selected technique(es))}

\item{project}{- A string or a vector of string (name(s) of selected projects)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all recent experiments.

names of recent experiments) or ('error', error_message)
}
\seealso{
Other Inserting and listing experiments: \code{\link{deepblue.add_experiment}},
  \code{\link{deepblue.faceting_experiments}},
  \code{\link{deepblue.list_experiments}},
  \code{\link{deepblue.list_similar_experiments}}
}

