% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_similar_biosources}
\alias{deepblue.list_similar_biosources}
\title{list_similar_biosources}
\usage{
deepblue.list_similar_biosources(name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (biosource name)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all biosources similar to the one provided.

similar biosources) or ('error', error_message)
}
\seealso{
Other Inserting and listing biosources: \code{\link{deepblue.add_biosource}},
  \code{\link{deepblue.list_biosources}}
}

