% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_similar_projects}
\alias{deepblue.list_similar_projects}
\title{list_similar_projects}
\usage{
deepblue.list_similar_projects(name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (project name)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all projects similar to the one provided.

similar project names) or ('error', error_message)
}
\seealso{
Other Inserting and listing projects: \code{\link{deepblue.add_project}},
  \code{\link{deepblue.add_user_to_project}},
  \code{\link{deepblue.list_projects}},
  \code{\link{deepblue.set_project_public}}
}

