% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.list_similar_techniques}
\alias{deepblue.list_similar_techniques}
\title{list_similar_techniques}
\usage{
deepblue.list_similar_techniques(name = NULL, user_key = deepblue.USER_KEY)
}
\arguments{
\item{name}{- A string (technique name)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Lists all techniques similar to the one provided.

similar techniques) or ('error', error_message)
}
\seealso{
Other Inserting and listing techniques: \code{\link{deepblue.add_technique}},
  \code{\link{deepblue.list_techniques}}
}

