% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepblue.r
\name{deepblue.score_matrix}
\alias{deepblue.score_matrix}
\title{score_matrix}
\usage{
deepblue.score_matrix(experiments_columns = NULL,
  aggregation_function = NULL, aggregation_regions_id = NULL,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{experiments_columns}{- A struct (map with experiments names and columns to be processed. Example : {'wgEncodeBroadHistoneDnd41H3k27acSig.wig':'VALUE', 'wgEncodeBroadHistoneCd20ro01794H3k27acSig.wig':'VALUE'})}

\item{aggregation_function}{- A string (aggregation function name: min, max, mean, var, sd, median, count, boolean)}

\item{aggregation_regions_id}{- A string (query ID of the regions that will be used as the aggregation boundaries)}

\item{user_key}{- A string (users token key)}
}
\value{
a vector with  ('okay',
}
\description{
Build a matrix containing the aggregation result of the the experiments data by aggregation regions.

BED formated regions) or ('error', error_message)
}
\seealso{
Other Operating on the data regions: \code{\link{deepblue.aggregate}},
  \code{\link{deepblue.count_regions}},
  \code{\link{deepblue.filter_regions}},
  \code{\link{deepblue.flank}},
  \code{\link{deepblue.get_experiments_by_query}},
  \code{\link{deepblue.get_regions}},
  \code{\link{deepblue.input_regions}},
  \code{\link{deepblue.intersection}},
  \code{\link{deepblue.merge_queries}},
  \code{\link{deepblue.query_cache}},
  \code{\link{deepblue.query_experiment_type}},
  \code{\link{deepblue.select_annotations}},
  \code{\link{deepblue.select_experiments}},
  \code{\link{deepblue.select_regions}},
  \code{\link{deepblue.tiling_regions}}
}

