% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpingFunctions.R
\name{process_request}
\alias{process_request}
\title{process_request}
\usage{
process_request(requested_regions, sleep.time = 1,
  user_key = deepblue.USER_KEY)
}
\arguments{
\item{requested_regions}{A string}

\item{sleep.time}{An integer with default value 1s}

\item{user_key}{A string}
}
\description{
Process the user request. Takes in three parameters; requested regions, sleep time, and user key.
}

